% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_file.R
\name{upload_sharepoint_file}
\alias{upload_sharepoint_file}
\title{Upload SharePoint File}
\usage{
upload_sharepoint_file(src, site_url, dest_fldr_url, token)
}
\arguments{
\item{src}{The file path to the local file that is being uploaded to
SharePoint}

\item{site_url}{This is the url when first accessing the sharepoint team via
your web browser (Home page)}

\item{dest_fldr_url}{This is a direct link to the SharePoint folder where the
file will be uploaded to. It can be obtained by navigating to the folder you
want to upload to, right-clicking and selecting "Copy Link"}

\item{token}{The token generated with the correct SharePoint permissions. Use
get_az_tk('sp") to create this token.}
}
\value{
Upload local file to selected SharePoint folder.
}
\description{
Upload SharePoint File
}
\examples{
\dontrun{
# Create SharePoint azure token
tk <- get_az_tk('sp')

# Example to upload local file to SharePoint folder
upload_sharepoint_file(
  src = 'TestFiles/test.txt',
  site_url = "https://myexample.sharepoint.com/sites/Example-Files/"),
  file_url = paste0("https://myexample.sharepoint.com/:x:/r/sites/Example-Files/Shared\%20Documents/Example_datasets/example_data_amended.csv?d=XXXXXXXXXXXXXXXXXXXXXXXXXXXX"),
  token = tk
)
}
}
